moving clim/ref functions to utils/utils
