+++ .gitignore
@@ -13,6 +13,7 @@ ecmean.egg-info/
 ecmean/utils/history_benchmark
 dist/
 pluto/
+coverage.lcov
 *.ipynb
